Created the word-count dictonary file
Program reads in entire textfile, scrubs the text, then splits the text into a list of strings (words), then the word frequencies are counted and a set is created with each pair containing a word along with the number of occurances of that word. The set is then sorted in alphabetically descending order and written to the given output file.